refactor(manager): extract ProviderManager from RadarProductManager
Move refresh scheduling and provider lifecycle into dedicated files as
the first step toward the multi-provider design in discussion #506.
No behavior change; adds a name-formatting unit test and small guards
for null provider and duplicate Shutdown.
