+++ test/source/scwx/qt/manager/radar_product_manager.test.cpp
@@ -1,4 +1,7 @@
 #include <scwx/qt/manager/radar_product_manager.hpp>
+// Include after radar_product_manager.hpp: Qt's emit macro breaks TBB if
+// reversed.
+#include <scwx/qt/manager/provider_manager.hpp>
 #include <scwx/qt/config/radar_site.hpp>
 #include <scwx/qt/util/geographic_lib.hpp>
 #include <scwx/common/constants.hpp>
@@ -107,6 +110,27 @@ TEST(RadarProductManager, CoordinateGenerationMatchesGeodesicReference)
                              units::angle::degrees<float> {1.0f},
                              units::angle::degrees<float> {0.5f},
                              0.5f);
+}
+
+TEST(ProviderManager, NameFormatting)
+{
+   config::RadarSite::Initialize();
+
+   RadarProductManager radarProductManager {"KLSX"};
+   ProviderManager     level2ProviderManager {
+      &radarProductManager, "KLSX", common::RadarProductGroup::Level2};
+   ProviderManager level3ProviderManager {
+      &radarProductManager, "KLSX", common::RadarProductGroup::Level3, "N0B"};
+   ProviderManager level2ChunksProviderManager {
+      &radarProductManager,
+      "KLSX",
+      common::RadarProductGroup::Level2,
+      "???",
+      true};
+
+   EXPECT_EQ(level2ProviderManager.name(), "KLSX, L2");
+   EXPECT_EQ(level3ProviderManager.name(), "KLSX, L3, N0B");
+   EXPECT_EQ(level2ChunksProviderManager.name(), "KLSX, L2");
 }
 
 } // namespace scwx::qt::manager
